Brazilian Bryophyte Taxonomy

A dataset containing the most relevant taxonomic information of Bryophyta stored in the Brazilian Flora 2020 taxonomic backbone, a.k.a. the Flora and Funga of Brazil. It includes all taxonomic levels (i.e. infra-species, species, genus, family and so on).

The original backbone was slightly edited aiming to standardize the notation across all backbones in plantRdata. Many columns in the original source that were not essential for the process of taxon name checking were removed and most of them were renamed. Duplicated scientific names (i.e. taxon name + author) whose taxon status were not accepted were also removed to avoid possible problems in the taxon name matching process.

data(bfoNamesBryophyta)

Format

An object of class data.frame with 14 columns and about 5 thousand rows.

Source

https://ipt.jbrj.gov.br/jbrj/resource?r=lista_especies_flora_brasil

Note

Last update/change of the downloaded backbone (day month year): 01 Oct 2025.

References

Flora e Funga do Brasil (Constantly updated): Flora e Funga do Brasil project - Projeto Flora e Funga do Brasil. Instituto de Pesquisas Jardim Botânico do Rio de Janeiro. Dataset/Checklist. doi:10.15468/1mtkaw.